+++ spec/www/spec/db-open-close-delete-test.js
@@ -109,7 +109,7 @@ var mytests = function() {
                 expect(rs.rows.length).toBe(1);
                 expect(rs.rows.item(0).name).toBe('main');
                 expect(rs.rows.item(0).file).toBeDefined();
-                expect(rs.rows.item(0).file.indexOf(dbName)).not.toBe(-1);
+                expect(rs.rows.item(0).file.indexOf(dbName) !== -1).toBe(true);
 
                 // Close & finish:
                 db.close(done, done);
@@ -147,7 +147,7 @@ var mytests = function() {
                 expect(rs.rows.length).toBe(1);
                 expect(rs.rows.item(0).name).toBe('main');
                 expect(rs.rows.item(0).file).toBeDefined();
-                expect(rs.rows.item(0).file.indexOf(dbName)).not.toBe(-1);
+                expect(rs.rows.item(0).file.indexOf(dbName) !== -1).toBe(true);
 
                 // Close & finish:
                 db.close(done, done);
@@ -211,7 +211,7 @@ var mytests = function() {
                 expect(rs.rows.length).toBe(1);
                 expect(rs.rows.item(0).name).toBe('main');
                 expect(rs.rows.item(0).file).toBeDefined();
-                expect(rs.rows.item(0).file.indexOf(dbName)).not.toBe(-1);
+                expect(rs.rows.item(0).file.indexOf(dbName) !== -1).toBe(true);
 
                 // Close & finish:
                 db.close(done, done);
@@ -255,7 +255,7 @@ var mytests = function() {
                 expect(rs.rows.length).toBe(1);
                 expect(rs.rows.item(0).name).toBe('main');
                 expect(rs.rows.item(0).file).toBeDefined();
-                expect(rs.rows.item(0).file.indexOf(dbName)).not.toBe(-1);
+                expect(rs.rows.item(0).file.indexOf(dbName) !== -1).toBe(true);
 
                 // Close & finish:
                 db.close(done, done);
@@ -295,7 +295,7 @@ var mytests = function() {
                 expect(rs.rows.length).toBe(1);
                 expect(rs.rows.item(0).name).toBe('main');
                 expect(rs.rows.item(0).file).toBeDefined();
-                expect(rs.rows.item(0).file.indexOf(dbName)).not.toBe(-1);
+                expect(rs.rows.item(0).file.indexOf(dbName) !== -1).toBe(true);
 
                 // Close & finish:
                 db.close(done, done);
@@ -320,8 +320,6 @@ var mytests = function() {
         }, MYTIMEOUT);
 
         it(suiteName + 'Open database with U+0801 (3-byte Samaritan character Bit) & check internal database file name', function(done) {
-          // if (!isWindows && isAndroid && !isImpl2) pending('XXX CRASH on Android (default evcore-native-driver database access implementation)');
-
           var dbName = 'a\u0801.db';
 
           try {
@@ -335,7 +333,7 @@ var mytests = function() {
                 expect(rs.rows.length).toBe(1);
                 expect(rs.rows.item(0).name).toBe('main');
                 expect(rs.rows.item(0).file).toBeDefined();
-                expect(rs.rows.item(0).file.indexOf(dbName)).not.toBe(-1);
+                expect(rs.rows.item(0).file.indexOf(dbName) !== -1).toBe(true);
 
                 // Close & finish:
                 db.close(done, done);
@@ -360,7 +358,7 @@ var mytests = function() {
         }, MYTIMEOUT);
 
         it(suiteName + 'Open database with emoji \uD83D\uDE03 (UTF-8 4 bytes) & check internal database file name', function(done) {
-          if (!isWindows && isAndroid && !isImpl2) pending('XXX BROKEN on Android (default evcore-native-driver database access implementation)'); // XXX
+          if (!isWindows && isAndroid && !isImpl2) pending('TBD KNOWN ISSUE on Android (default evcore-native-driver database access implementation)');
 
           var dbName = 'a\uD83D\uDE03.db';
 
@@ -375,7 +373,7 @@ var mytests = function() {
                 expect(rs.rows.length).toBe(1);
                 expect(rs.rows.item(0).name).toBe('main');
                 expect(rs.rows.item(0).file).toBeDefined();
-                expect(rs.rows.item(0).file.indexOf(dbName)).not.toBe(-1);
+                expect(rs.rows.item(0).file.indexOf(dbName) !== -1).toBe(true);
 
                 // Close & finish:
                 db.close(done, done);
@@ -440,7 +438,7 @@ var mytests = function() {
                   expect(rs.rows.length).toBe(1);
                   expect(rs.rows.item(0).name).toBe('main');
                   expect(rs.rows.item(0).file).toBeDefined();
-                  expect(rs.rows.item(0).file.indexOf(dbName)).not.toBe(-1);
+                  expect(rs.rows.item(0).file.indexOf(dbName) !== -1).toBe(true);
 
                   // Close & finish:
                   db.close(done, done);
@@ -501,7 +499,7 @@ var mytests = function() {
                   expect(rs.rows.length).toBe(1);
                   expect(rs.rows.item(0).name).toBe('main');
                   expect(rs.rows.item(0).file).toBeDefined();
-                  expect(rs.rows.item(0).file.indexOf(dbName)).not.toBe(-1);
+                  expect(rs.rows.item(0).file.indexOf(dbName) !== -1).toBe(true);
 
                   // Close & finish:
                   db.close(done, done);
@@ -536,6 +534,9 @@ var mytests = function() {
       });
     }
 
+    describe('Web SQL vs plugin openDatabase test(s)', function() {
+
+      if (true) {
 
         it('Open plugin database with Web SQL parameters - REJECTED AGAIN in this plugin version', function(done) {
           try {
@@ -582,7 +583,10 @@ var mytests = function() {
           }
         }, MYTIMEOUT);
 
-      if (window.hasWebKitWebSQL)
+      }
+
+      if (window.hasWebKitWebSQL) {
+
         it('Web SQL check that db name is really a string', function(done) {
           var p1 = { name: 'my.db.name', location: 'default' };
           try {
@@ -603,9 +607,13 @@ var mytests = function() {
           }
         }, MYTIMEOUT);
 
-    for (var i=0; i<pluginScenarioCount; ++i) {
+      }
 
-      describe(pluginScenarioList[i] + ': open database location parameter test(s)', function() {
+    });
+
+    describe('Plugin: open database location parameter test(s)', function() {
+
+      if (true) {
 
         var suiteName = 'Plugin: ';
 
@@ -924,9 +932,62 @@ var mytests = function() {
           }
         }, MYTIMEOUT);
 
-      });
+        it(suiteName + 'open with androidDatabaseProvider: "bogus" - REJECTED with exception', function(done) {
+          try {
+            window.sqlitePlugin.openDatabase({
+              name: 'open-with-androidDatabaseProvider-bogus.db',
+              androidDatabaseProvider: 'bogus',
+              location: 'default'
+            }, function(db) {
+              // NOT EXPECTED:
+              expect(false).toBe(true);
 
-    }
+              // Close (plugin) & finish:
+              db.close(done, done);
+            }, function(error) {
+              // OK but NOT EXPECTED:
+              expect('Behavior changed, please update this test').toBe('--');
+
+              done();
+            });
+          } catch (e) {
+            // EXPECTED RESULT: stopped by the implementation
+            expect(e).toBeDefined();
+
+            done();
+          }
+        }, MYTIMEOUT);
+
+        it(suiteName + 'open with both androidDatabaseProvider and androidDatabaseImplementation - REJECTED with exception', function(done) {
+          try {
+            window.sqlitePlugin.openDatabase({
+              name: 'open-with-androidDatabaseProvider-and-androidDatabaseImplementation.db',
+              androidDatabaseProvider: 'system',
+              androidDatabaseImplementation: 2,
+              location: 'default'
+            }, function(db) {
+              // NOT EXPECTED:
+              expect(false).toBe(true);
+
+              // Close (plugin) & finish:
+              db.close(done, done);
+            }, function(error) {
+              // OK but NOT EXPECTED:
+              expect('Behavior changed, please update this test').toBe('--');
+
+              done();
+            });
+          } catch (e) {
+            // EXPECTED RESULT: stopped by the implementation
+            expect(e).toBeDefined();
+
+            done();
+          }
+        }, MYTIMEOUT);
+
+      }
+
+    });
 
   });
 
@@ -935,7 +996,11 @@ var mytests = function() {
     // TBD skip plugin test on browser platform (not yet supported):
     if (isBrowser) return;
 
+    if (true) {
+
       var suiteName = 'plugin: ';
+
+      if (true) {
 
         it(suiteName + 'check that sqlitePlugin.deleteDatabase db name is really a string', function(done) {
           var p1 = { name: 'my.db.name', location: 1 };
@@ -1234,9 +1299,13 @@ var mytests = function() {
           }
         }, MYTIMEOUT);
 
+      }
+
+    }
+
   });
 
-  describe('Plugin: db open-close-delete test(s)', function() {
+  describe('Plugin: db open-close-delete operation behavior test(s)', function() {
     // TBD skip plugin test on browser platform (not yet supported):
     if (isBrowser) return;
 
@@ -1486,7 +1555,7 @@ var mytests = function() {
 
         test_it(suiteName + ' database.close (immediately after open) calls its success callback', function () {
           // TBD POSSIBLY BROKEN on iOS/macOS due to current background processing implementation:
-          if (!isAndroid && !isWindows) pending('CURRENTLY BROKEN on iOS/macOS (background processing implementation)');
+          if (isAppleMobileOS || isMac) pending('TBD POSSIBLY BROKEN on iOS/macOS (background processing implementation)');
 
           // asynch test coming up
           stop(1);
@@ -1689,7 +1758,7 @@ var mytests = function() {
           window.sqlitePlugin.deleteDatabase(first, second, third);
         }
 
-        // Needed to support some large-scale applications:
+        // (Was needed to test support for an enterprise application)
         test_it(suiteName + ' open same database twice in [same] specified location works', function () {
           // XXX TODO [BROKEN]: same db name, different location should be different db!
           stop(2);
@@ -1727,7 +1796,7 @@ var mytests = function() {
           });
         });
 
-        // Needed to support some large-scale applications:
+        // (Was needed to test support for an enterprise application)
         test_it(suiteName + ' close then re-open (2x) allows subsequent queries to run', function () {
           // asynch test coming up
           stop(1);
@@ -1778,7 +1847,7 @@ var mytests = function() {
           });
         });
 
-        // Needed to support some large-scale applications:
+        // (Was needed to test support for an enterprise application)
         test_it(suiteName + " delete then re-open (location: 'default') allows subsequent queries to run", function () {
           var dbName = "test-database-delete-and-reopen.db";
           var dbargs = {name: dbName, iosDatabaseLocation: 'default'};
@@ -1820,9 +1889,9 @@ var mytests = function() {
         });
 
         // XXX SEE BELOW: repeat scenario but wait for open callback before close/delete/reopen
-        // Needed to support some large-scale applications:
+        // (Was needed to test support for an enterprise application)
         test_it(suiteName + ' immediate close, then delete then re-open allows subsequent queries to run', function () {
-          if (!isAndroid && !isWindows) pending('CURRENTLY BROKEN on iOS/macOS (background processing implementation)');
+          if (isAppleMobileOS || isMac) pending('CURRENTLY BROKEN on iOS/macOS (background processing implementation)');
 
           var dbName = "Immediate-close-delete-Reopen.db";
           var dbargs = {name: dbName, location: 'default'};
@@ -1969,7 +2038,7 @@ var mytests = function() {
 
         test_it(suiteName + ' repeatedly open and close database faster (5x)', function () {
           // TBD CURRENTLY BROKEN on iOS/macOS due to current background processing implementation:
-          if (!isAndroid && !isWindows) pending('CURRENTLY BROKEN on iOS/macOS (background processing implementation)');
+          if (isAppleMobileOS || isMac) pending('CURRENTLY BROKEN on iOS/macOS (background processing implementation)');
           // TBD ???:
           if (isAndroid && isImpl2) pending('FAILS on builtin android.database implementation (androidDatabaseImplementation: 2)');
 
@@ -2027,7 +2096,7 @@ var mytests = function() {
           });
         });
 
-        // Needed to support some large-scale applications:
+        // (Was needed to test support for an enterprise application)
         test_it(suiteName + ' repeatedly open and delete database (4x)', function () {
           var dbName = 'test-repeatedly-open-and-delete-4x.db';
           var dbargs = {name: dbName, iosDatabaseLocation: 'Documents'};
@@ -2090,7 +2159,7 @@ var mytests = function() {
           });
         });
 
-        // Needed to support some large-scale applications:
+        // (Was needed to test support for an enterprise application)
         test_it(suiteName + ' repeatedly open and delete database faster (5x)', function () {
           // TBD POSSIBLY BROKEN on iOS/macOS ...
           // if (!isAndroid && !isWindows) pending(...);