+++ CHANGES.md
@@ -1,5 +1,23 @@
 # Changes
 
+# cordova-sqlite-evplus-ext-free 0.3.7
+
+## cordova-sqlite-evcore-extbuild-free 0.9.10
+
+- Include SQLite 3.26.0 update from cordova-sqlite-evcore-free-dependencies 0.8.7
+
+#### cordova-sqlite-storage 2.5.1
+
+- fix internal plugin cleanup error log on Android
+
+#### cordova-sqlite-storage 2.5.0
+
+- androidDatabaseProvider: 'system' setting, to replace androidDatabaseImplementation setting which is now deprecated
+
+### cordova-sqlite-evcore-common-free 0.0.3
+
+- Internal error message fix for evcore on Android
+
 # cordova-sqlite-evplus-ext-free 0.3.6
 
 - evcore-native-driver.jar from cordova-sqlite-evcore-free-dependencies 0.8.6, with cleaner workaround for 4-byte UTF-8 crash bug and fix for Samaritan character crash bug on Android
@@ -101,7 +119,7 @@
 
 ## cordova-sqlite-evcore-extbuild-free 0.9.2
 
-### cordova-sqlite-storage 2.2.0
+#### cordova-sqlite-storage 2.2.0
 
 - Fix SQLiteAndroidDatabase implementation for Turkish and other foreign locales